% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for a given use-case}
\usage{
prep_data(use_case)
}
\arguments{
\item{use_case}{"hurricane" or "beauty".}
}
\value{
A list: (df, outcome_var, outcome_var_label, strip_vars, variable_labels)
}
\description{
Prepare data for a given use-case
}
\examples{
\dontrun{
prep <- prep_data("hurricane")
}
}
