% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_heatmap_strips.R
\name{generate_heatmap_strips}
\alias{generate_heatmap_strips}
\title{Heatmap strips for analytical decisions}
\usage{
generate_heatmap_strips(
  data,
  outcome_var,
  strip_vars,
  variable_labels,
  n_bins = 20,
  palette_option = "D"
)
}
\arguments{
\item{data}{Data frame (with factor decision vars).}

\item{outcome_var}{Name of outcome column.}

\item{strip_vars}{Character vector of factor decision variables.}

\item{variable_labels}{Named vector for display labels.}

\item{n_bins}{Number of x bins.}

\item{palette_option}{Viridis palette option.}
}
\value{
A patchwork object.
}
\description{
Heatmap strips for analytical decisions
}
\examples{
\dontrun{
prep <- prep_data("hurricane")
generate_heatmap_strips(prep[[1]], prep[[2]], prep[[4]], prep[[5]])
}
}
